+++ src/SwaggerProvider.DesignTime/v3/OperationCompiler.fs
@@ -8,7 +8,7 @@ open System.Text.RegularExpressions
 
 open Microsoft.FSharp.Quotations
 open Microsoft.FSharp.Quotations.ExprShape
-open Microsoft.OpenApi.Models
+open Microsoft.OpenApi
 open ProviderImplementation.ProvidedTypes
 open FSharp.Data.Runtime.NameUtils
 
@@ -19,7 +19,7 @@ open Swagger.Internal
 // We cannot use record here
 // TP cannot load DTC with OpenApiPathItem/OperationType props (from 3rd party assembly)
 // Probably related to https://github.com/fsprojects/FSharp.TypeProviders.SDK/issues/274
-type ApiCall = string * OpenApiPathItem * OperationType
+type ApiCall = string * IOpenApiPathItem * HttpMethod
 
 [<Struct>]
 type PayloadType =
@@ -67,29 +67,39 @@ type OperationCompiler(schema: OpenApiDocument, defCompiler: DefinitionCompiler,
         if String.IsNullOrWhiteSpace providedMethodName then
             failwithf $"Operation name could not be empty. See '%s{path}/%A{opTy}'"
 
-        let unambiguousName(par: OpenApiParameter) =
+        let unambiguousName(par: IOpenApiParameter) =
             $"%s{par.Name}In%A{par.In}"
 
-        let openApiParameters = [ yield! pathItem.Parameters; yield! operation.Parameters ]
+        let openApiParameters =
+            [ if not(isNull pathItem.Parameters) then
+                  yield! pathItem.Parameters
+              if not(isNull operation.Parameters) then
+                  yield! operation.Parameters ]
 
         let (|MediaType|_|) contentType (content: IDictionary<string, OpenApiMediaType>) =
-            match content.TryGetValue contentType with
-            | true, mediaTyObj -> Some(mediaTyObj)
-            | _ -> None
+            if isNull content then
+                None
+            else
+                match content.TryGetValue contentType with
+                | true, mediaTyObj -> Some mediaTyObj
+                | _ -> None
 
         let (|TextReturn|_|)(input: string) =
             if input.StartsWith("text/") then Some(input) else None
 
         let (|TextMediaType|_|)(content: IDictionary<string, OpenApiMediaType>) =
-            content.Keys |> Seq.tryPick (|TextReturn|_|)
+            if isNull content then
+                None
+            else
+                content.Keys |> Seq.tryPick (|TextReturn|_|)
 
         let (|NoMediaType|_|)(content: IDictionary<string, OpenApiMediaType>) =
-            if content.Count = 0 then Some() else None
+            if isNull content || content.Count = 0 then Some() else None
 
         let payloadMime, parameters =
             /// handles de-duplicating Swagger parameter names if the same parameter name
             /// appears in multiple locations in a given operation definition.
-            let uniqueParamName usedNames (param: OpenApiParameter) =
+            let uniqueParamName usedNames (param: IOpenApiParameter) =
                 let name = niceCamelName param.Name
 
                 if usedNames |> Set.contains name then
@@ -110,6 +120,7 @@ type OperationCompiler(schema: OpenApiDocument, defCompiler: DefinitionCompiler,
                                 Schema = schema,
                                 Required = true //operation.RequestBody.Required
                             )
+                            :> IOpenApiParameter
 
                         Some(payloadType, p)
 
@@ -142,7 +153,7 @@ type OperationCompiler(schema: OpenApiDocument, defCompiler: DefinitionCompiler,
                       if bodyFormatAndParam.IsSome then
                           yield bodyFormatAndParam.Value |> snd ]
                     |> List.distinctBy(fun op -> op.Name, op.In)
-                    |> List.partition(fun x -> x.Required)
+                    |> List.partition(_.Required)
 
                 List.append required optional
 
@@ -285,7 +296,7 @@ type OperationCompiler(schema: OpenApiDocument, defCompiler: DefinitionCompiler,
                         let path, queryParams, headers =
                             let path, queryParams, headers, cookies =
                                 ((<@ path @>, <@ [] @>, headers, <@ [] @>), parameters)
-                                ||> List.fold(fun (path, query, headers, cookies) (param: OpenApiParameter, valueExpr) ->
+                                ||> List.fold(fun (path, query, headers, cookies) (param: IOpenApiParameter, valueExpr) ->
                                     if param.In.HasValue then
                                         let name = param.Name
 
@@ -481,12 +492,16 @@ type OperationCompiler(schema: OpenApiDocument, defCompiler: DefinitionCompiler,
 
         List.ofSeq schema.Paths
         |> List.collect(fun path ->
-            if path.Value.UnresolvedReference then
-                failwith
-                    $"TP does not support unresolved paths / external references. Path '%s{path.Key}' refer to '%s{path.Value.Reference.ReferenceV3}'"
+            // if path.Value.UnresolvedReference then
+            //     failwith
+            //         $"TP does not support unresolved paths / external references. Path '%s{path.Key}' refer to '%s{path.Value.Reference.ReferenceV3}'"
 
-            List.ofSeq path.Value.Operations
+            let safeSeq s =
+                if isNull s then Seq.empty else s
+
+            List.ofSeq(safeSeq path.Value.Operations)
             |> List.map(fun kv -> path.Key, path.Value, kv.Key))
+
         |> List.groupBy(fun (_, pathItem, opTy) ->
             if ignoreControllerPrefix then
                 String.Empty //