+++ docs/RELEASE_NOTES.md
@@ -1,8 +1,11 @@
-#### 3.0.0-beta02 - Dec 22, 2025
+#### 3.0.0-beta03 - Dec 24, 2025
 
-- Server-Side Request Forgery (SSRF) protection for IP V4 (#271) and V6 (#272)
+- feat: Microsoft.OpenApi v2.x with OpenAPI v3.1 Support
+- feat: Server-Side Request Forgery (SSRF) protection for IP V4 (#271) and V6 (#272)
 - feat: handle text/plain payload types (#276)
+- fix: removed dependency on Microsoft.OpenApi.Readers
 - fix: error messages when `operationId` is not provided (minimal api) (#277)
+- hk: dependencies update
 
 #### 2.3.2 - Oct 11, 2025